backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
qdrant-client>=1.10.0
asyncpg>=0.27.0
psycopg[binary]>=3.1.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0